+++ docs/changelog.rst
@@ -17,6 +17,11 @@ Release Notes
 To avail of fixes in an unreleased version, please download a ZIP file
 `directly from GitHub <https://github.com/mitogen-hq/mitogen/>`_.
 
+
+Unreleased
+----------
+
+
 v0.3.6 (2024-04-04)
 -------------------